data/figures: Extract figures from the CSS file to keep sanity.
These data files should not be part of the CSS (which is supposed to be
"source code") and, if possible, should be sourced from external files.

If we are really looking to squeeze the last bytes (which we weren't, since
base64 encoded files are usually 4/3 the original size), we can consider the
option of using sprites for this purpose, but that's a decision for the
future.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>