data: figures: Optimize aggressively the PNG files.
I optimized (losslessly) the PNG files with the following sequence of
commands:

    optipng -zc1-9 -zm1-9 -zs0-3 -f0-5 *.png
    advpng -z3 *.png
    advpng -z4 -i2500 *.png
    advdef -z3 *.png
    advdef -z4 -i2500 *.png

and the savings were amazing in relation with what we had before: about 10KB
(which, and if we convert back with base64 encoding, would result in 4/3 *
10 = 13 fewer KB).

This goes on to show that the previous strategy of mixing the code with data
renders both unreadable code as well as hindering software maintenance.
Furthermore, any review process might consider this an improvement, as there
are no blobs embedded in the code (once we remove them from the CSS files).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>