data: First try at introducing a dark mode for reading.
This dark mode is based on the original Athelas mode from Readability (with
the same colors), but, of course, without font blobs. Furthermore, only Free
fonts are used.

Note that the icon for the dark mode is currently the same as the one for
terminal mode, but once things get working, this is simply a matter of
cosmetics.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>